==> 3_goedicke_plausibilisierung.qmd <==
---
title: "goedicke_plausibilisierung"
format: html
---

Load necessary libraries:
```{r}
library(readr)
library(dplyr)
library(nVennR)
library(VennDiagram)
library(ggplot2)
library(stringr)
source("library.R")
```

Read combined data:
```{r}
data <- read_csv("../Daten/Intermediate/combined_data.csv")
data <- data[data$year != "goedicke" | data$in_vicinity,]
```


Compute diff between goedicke and climax and check whether species present in climax and not in goedicke are present on the fire patch:
```{r}
unique_species <- function(species) {
  all_fully_determined_species <- unique(species[!endsWith(species, "spec.")])
  all_spec_genus <- unique(first_word(species[endsWith(species, "spec.")]))
  all_fully_determined_genus <- unique(first_word(species[!endsWith(species, "spec.")]))
  all_additional_spec_species <- all_spec_genus[!(all_spec_genus %in% all_fully_determined_genus)]
  result <- all_fully_determined_species
  if(length(all_additional_spec_species) > 0) {
    result <- append(all_fully_determined_species, paste(all_additional_spec_species, "spec."))
  }
  return(result)
}

title <- "all altitudes"
analyze_seed_backpressure <- function(venn_data, title) {
  length(unique(venn_data$global_id[venn_data$year == "goedicke"]))
  length(unique(venn_data$global_id[venn_data$year == "climax"]))
  
  goedicke_species <- unique_species(venn_data$`Species_unified_2004-23`[venn_data$year == "goedicke"])
  goedicke_species
  
  climax_species <- unique_species(venn_data$`Species_unified_2004-23`[venn_data$year == "climax"])
  climax_species
  
  fire_species <- unique_species(venn_data$`Species_unified_2004-23`[venn_data$year != "goedicke" & venn_data$year != "climax"])
  fire_species
  
  ncnfg <- length(setdiff(setdiff(goedicke_species, climax_species), fire_species))
  cnfg <- length(intersect(setdiff(climax_species, fire_species), goedicke_species))
  cnfng <- length(setdiff(setdiff(climax_species, fire_species), goedicke_species))
  ncfg <- length(setdiff(intersect(goedicke_species, fire_species), climax_species))
  cfg <- length(intersect(intersect(climax_species, fire_species), goedicke_species))
  cfng <- length(setdiff(intersect(climax_species, fire_species), goedicke_species))
  ncfng <- length(setdiff(setdiff(fire_species, climax_species), goedicke_species))

  # there is still a difference...
  ncnfg + cnfg + cnfng + cfng + cfg + ncfg + ncfng
  length(unique(venn_data$`Species_unified_2004-23`))

  #create dataset to plot
  year <- c(rep("goedicke", 3), rep("climax", 3))
  present <- rep(c("unique", "fire", "common"))
  number_of_species <- c(ncnfg, ncfg, cfg, cnfng, cfng, cfg)
  plot_data <- data.frame(year,present,number_of_species)
  ggplot(plot_data, aes(fill=present, y=number_of_species, x=year)) + 
      geom_bar(position="stack", stat="identity")
  
  venn.diagram(
    x = list(fire_species, goedicke_species, climax_species),
    category.names = c("2004-2023", "Before", "Intact"),
    filename = paste("../Daten/Results/Goedicke/", title, "simple.svg"),
    output = T
  )
  
  species_venn <- createVennObj(nSets = 3, sNames = c('Before', '2004-2023', 'Intact'))
  species_venn <- setVennRegion(species_venn, c('2004-2023'), as.numeric(ncfng))
  species_venn <- setVennRegion(species_venn, c('Before'), as.numeric(ncnfg))
  species_venn <- setVennRegion(species_venn, c('Intact'), as.numeric(cnfng))
  species_venn <- setVennRegion(species_venn, c('2004-2023', 'Before', 'Intact'), as.numeric(cfg))
  species_venn <- setVennRegion(species_venn, c('2004-2023', 'Before'), as.numeric(ncfg))
  species_venn <- setVennRegion(species_venn, c('2004-2023', 'Intact'), as.numeric(cfng))
  species_venn <- setVennRegion(species_venn, c('Before', 'Intact'), as.numeric(cnfg))
  species_venn <- plotVenn(nVennObj = species_venn)
  showSVG(species_venn, opacity=0.2, outFile = paste("../Daten/Results/Goedicke/", title, ".svg"))
  
  
  all_species <- unique(venn_data$`Species_unified_2004-23`)
  
  percent_plots_containing_species <- data.frame(species = character(), goedicke = numeric(), fire = numeric(), climax = numeric())
  
  species <- all_species[1]
  for(species in all_species) {
    percent_goedicke_plots <- 100 * length(unique(venn_data$global_id[venn_data$`Species_unified_2004-23` == species & venn_data$year == "goedicke"])) / length(unique(venn_data$global_id[venn_data$year == "goedicke"]))
    percent_fire_plots <- 100 * sum((venn_data[!venn_data$year %in% c("goedicke", "climax"),] |> group_by(global_id, year) |> summarize(present = as.integer(sum((species %in% `Species_unified_2004-23`)) > 0)))$present) / nrow(venn_data[!venn_data$year %in% c("goedicke", "climax"),] |> group_by(global_id, year) |> summarize(n = 1))
    percent_climax_plots <- 100 * length(unique(venn_data$global_id[venn_data$`Species_unified_2004-23` == species & venn_data$year == "climax"])) / length(unique(venn_data$global_id[venn_data$year == "climax"]))
    
    percent_plots_containing_species <- add_row(percent_plots_containing_species, species = species, goedicke = percent_goedicke_plots, fire = percent_fire_plots, climax = percent_climax_plots)
  }
  
  percent_plots_containing_species
  percent_plots_containing_species$delta_fire <- percent_plots_containing_species$fire - percent_plots_containing_species$goedicke
  percent_plots_containing_species$delta_climax <- percent_plots_containing_species$climax - percent_plots_containing_species$goedicke
  
  cor(percent_plots_containing_species$delta_fire, percent_plots_containing_species$delta_climax, method = "pearson")
  res <- cor.test(percent_plots_containing_species$delta_fire, percent_plots_containing_species$delta_climax, alternative = "greater", method = "pearson")
  res$p.value
  
  sum <- summary(lm(delta_fire ~ delta_climax, percent_plots_containing_species))
  
  p_correlation_delta_precent <- 
    percent_plots_containing_species |> 
    ggplot(aes(x = delta_fire, y = delta_climax)) + 
      geom_point() + 
      geom_smooth(method = lm) + 
      ggtitle(title) +
      annotate("text", x = 30, y = 40, label = paste("pearson p-value = ", round(res$p.value, 3)))

  ggsave(filename = paste("../Daten/Results/Goedicke/", title, "delta_correlation.png"), plot = p_correlation_delta_precent)
  
  p_correlation_delta_precent_with_labels <- 
    percent_plots_containing_species |> 
    ggplot(aes(x = delta_fire, y = delta_climax)) + 
      geom_point() + 
      geom_smooth(method = lm) + 
      geom_label(label = percent_plots_containing_species$species, size = 2) + 
      ggtitle(title)
  
  ggsave(filename = paste("../Daten/Results/Goedicke/", title, "delta_correlation_with_labels.png"), plot = p_correlation_delta_precent_with_labels)
}

# only keep species which are contributing to the overall gamma diversity
venn_data <- data[data$`Species_unified_2004-23` %in% unique_species(data$`Species_unified_2004-23`),]

analyze_seed_backpressure(venn_data, "all altitudes")
analyze_seed_backpressure(venn_data[venn_data$altitude_band == "high",], "high altitudes")
analyze_seed_backpressure(venn_data[venn_data$altitude_band == "mid",], "mid altitudes")
analyze_seed_backpressure(venn_data[venn_data$altitude_band == "low",], "low altitudes")
```
