==> 5_red_list_neophyte_functional_group.qmd <==
---
title: "neophyte and red list analysis"
format: html
---

Load necessary libraries:
```{r}
library(readr)
library(dplyr)
library(ggplot2)
library(tidyr)
```

Read combined data:
```{r}
selected_years = c("goedicke", "2004", "2005", "2006", "2007", "2013", "2023", "climax")
# selected_years = c("goedicke", "2004", "2007", "2013", "2023", "climax")

labels=data.frame(years = c("goedicke", "2004", "2005", "2006", "2007", "2013", "2023", "climax"), labels = c("Before", "2004", "2005", "2006", "2007", "2013", "2023", "Intact"), short_labels = c("Before", "4", "5", "6", "7", "13", "23", "Intact"))

data <- read_csv("../Daten/Intermediate/combined_data.csv")
data$`Species_unified_2004-23` <- sub(" cf\\.", "", data$`Species_unified_2004-23`)
data <- data[data$year != "goedicke" | data$in_vicinity,] # remove goedicke plots which are far away

data$year <- factor(data$year, levels = c("goedicke", "2004", "2005", "2006", "2007", "2013", "2023", "climax"))
taxon_conversion <- read_csv("../Daten/Intermediate/leuk_landolt_try_midolo_eive_taxon_conversion.csv")
landolt_data <- read_csv("../Daten/Landolt/2023-03-29_FI-export_Stefan_Widmer.csv")

check_list <- read_csv("../Daten/InfoFlora/Checklist_2017_simple_version_20230503.csv")
check_list$neophyt <- !is.na(check_list$`Indigenat CH`) & check_list$`Indigenat CH` %in% c("N", "NA", "NC")

red_list <- read_csv("../Daten/RedList/RL_Arten_2016.csv")

temp <- left_join(data, taxon_conversion, join_by(`Species_unified_2004-23` == infoflora_taxon))
complete_data <- left_join(temp, landolt_data, join_by(landolt_taxa == Taxon))
complete_data <- left_join(complete_data, check_list, join_by(check_list == Taxonname))
complete_data <- left_join(complete_data, red_list, join_by(`Species_unified_2004-23` == `Scientific name`))
```

```{r}
unique(complete_data$`Species_unified_2004-23`[is.na(complete_data$EG)])
data_functional_group <- complete_data[!is.na(complete_data$EG),]
data_functional_group <- data_functional_group[data_functional_group$EG != '-',]
nrow(data_functional_group) / sum(data_functional_group$EG %in% c('1', '2', '3', '4', '5', '6', '7', '8'))


title <- "all altitudes"
data <- data_functional_group


compute_fraction_of_plants_in_functional_groups <- function(data) {
  data_functional_group <- distinct(data, year, global_id, `Species_unified_2004-23`, EG)

  result <- data_functional_group[data_functional_group$year %in% selected_years,] |> 
    dplyr::select(year, global_id, EG) |>
    group_by(year, global_id) |>
    summarise(
      p_forest_plant = 100 * sum(EG == '1') / length(global_id),
      p_mountain_plant = 100 * sum(EG == '2') / length(global_id),
      p_low_elevation_pioneer = 100 * sum(EG == '3') / length(global_id),
      p_water = 100 * sum(EG == '4') / length(global_id),
      p_wetland = 100 * sum(EG == '5') / length(global_id),
      p_dry_meadow = 100 * sum(EG == '6') / length(global_id),
      p_ruderal = 100 * sum(EG == '7') / length(global_id),
      p_fertilized_meadow = 100 * sum(EG == '8') / length(global_id)
    )
  return(result)
}

compute_number_of_plants_in_functional_groups <- function(data) {
  data_functional_group <- distinct(data, year, global_id, `Species_unified_2004-23`, EG)

  result <- data_functional_group[data_functional_group$year %in% selected_years,] |> 
    dplyr::select(year, global_id, EG) |>
    group_by(year, global_id) |>
    summarise(
      p_forest_plant = sum(EG == '1'),
      p_mountain_plant = sum(EG == '2'),
      p_low_elevation_pioneer = sum(EG == '3'),
      p_water = sum(EG == '4'),
      p_wetland = sum(EG == '5'),
      p_dry_meadow = sum(EG == '6'),
      p_ruderal = sum(EG == '7'),
      p_fertilized_meadow = sum(EG == '8')
    )
  return(result)
}

create_unequal_spaced_plot <- function(tmp, title) {
    tmp <- pivot_longer(tmp, c("p_forest_plant", "p_mountain_plant", "p_low_elevation_pioneer", "p_water", "p_wetland", "p_dry_meadow", "p_ruderal", "p_fertilized_meadow"))
  
  tmp <-  tmp |> 
    group_by(name, year) |>
    summarise(
      m = mean(value),
      se = sd(value) / sqrt(length(value))
    )
  
  tmp$x <- NA
  tmp$x[tmp$year == "goedicke"] <- 0
  tmp$x[tmp$year == "2004"] <- 13
  tmp$x[tmp$year == "2005"] <- 14
  tmp$x[tmp$year == "2006"] <- 15
  tmp$x[tmp$year == "2007"] <- 16
  tmp$x[tmp$year == "2013"] <- 22
  tmp$x[tmp$year == "2023"] <- 32
  tmp$x[tmp$year == "climax"] <- 42
  
  plot <- tmp |> 
    ggplot(aes(colour=name, group = name, y=m, x=x)) + 
      geom_errorbar(aes(ymin=m - se, ymax=m + se), width=1, position=position_dodge(0)) +
      geom_line(data = tmp[tmp$year %in% c("goedicke", "2004"),], linetype = 2) +
      geom_line(data = tmp[tmp$year %in% c("2004", "2005", "2006", "2007", "2013", "2023"),], linetype = 1) +
      geom_line(data = tmp[tmp$year %in% c("2023", "climax"),], linetype = 3) +
      geom_point() +
      scale_color_discrete(name = "Functional Group", labels = c("Dry Meadow", "Fertilized Meadow", "Forest", "Low Elevation Pioneer", "Mountain Plant", "Ruderal", "Water", "Wetland")) +
      scale_x_continuous(breaks = c(0,13,14,15,16,22,32,42), labels=labels$short_labels[labels$years %in% selected_years]) +
      ylab("Fraction of plants in Functional Group in %") +
      xlab("Year") +
      ggtitle(title) +
      theme_classic()
  return(plot)
}

create_equal_spaced_plot <- function(tmp, title) {
    tmp <- pivot_longer(tmp, c("p_forest_plant", "p_mountain_plant", "p_low_elevation_pioneer", "p_water", "p_wetland", "p_dry_meadow", "p_ruderal", "p_fertilized_meadow"))
  
  tmp <-  tmp |> 
    group_by(name, year) |>
    summarise(
      m = mean(value),
      se = sd(value) / sqrt(length(value))
    )
  
  plot <- tmp |> 
    ggplot(aes(colour=name, group = name, y=m, x=year)) + 
      geom_errorbar(aes(ymin=m - se, ymax=m + se), width=1, position=position_dodge(0)) +
      geom_line(data = tmp[tmp$year %in% c("goedicke", "2004"),], linetype = 2) +
      geom_line(data = tmp[tmp$year %in% c("2004", "2005", "2006", "2007", "2013", "2023"),], linetype = 1) +
      geom_line(data = tmp[tmp$year %in% c("2023", "climax"),], linetype = 3) +
      geom_point() +
      scale_color_discrete(name = "Functional Group", labels = c("Dry Meadow", "Fertilized Meadow", "Forest", "Low Elevation Pioneer", "Mountain Plant", "Ruderal", "Water", "Wetland")) +
      scale_x_discrete(labels=labels$labels[labels$years %in% selected_years]) +
      ylab("Fraction of plants in Functional Group in %") +
      xlab("Year") +
      ggtitle(title) +
      theme_classic()
  return(plot)
}

create_unequal_spaced_mean_fraction_of_plants_in_functional_groups_plot <- function(data, title) {
  plot <- create_unequal_spaced_plot(compute_fraction_of_plants_in_functional_groups(data), 
                                     paste("Mean fraction of plants in each functional group at ", title))
  ggsave(paste("../Daten/Results/FunctionalGroups/unequal_spaced_mean_fraction_of_plants_in_each_functional_group_at", title, ".png"), plot = plot, width = 2000, height = 1500, units = "px")
}

create_equal_spaced_mean_fraction_of_plants_in_functional_groups_plot <- function(data, title) {
  plot <- create_equal_spaced_plot(compute_fraction_of_plants_in_functional_groups(data), 
                                   paste("Mean fraction of plants in each functional group at ", title))
  ggsave(paste("../Daten/Results/FunctionalGroups/equal_spaced_mean_fraction_of_plants_in_each_functional_group_at", title, ".png"), plot = plot, width = 2000, height = 1500, units = "px")
}

create_unequal_spaced_mean_number_of_plants_in_functional_groups_plot <- function(data, title) {
  plot <- create_unequal_spaced_plot(compute_number_of_plants_in_functional_groups(data), 
                                     paste("Mean number of plants in each functional group at ", title))
  ggsave(paste("../Daten/Results/FunctionalGroups/unequal_spaced_mean_number_of_plants_in_each_functional_group_at", title, ".png"), plot = plot, width = 2000, height = 1500, units = "px")
}

create_equal_spaced_mean_number_of_plants_in_functional_groups_plot <- function(data, title) {
  plot <- create_equal_spaced_plot(compute_number_of_plants_in_functional_groups(data), 
                                   paste("Mean number of plants in each functional group at ", title))
  ggsave(paste("../Daten/Results/FunctionalGroups/equal_spaced_mean_number_of_plants_in_each_functional_group_at", title, ".png"), plot = plot, width = 2000, height = 1500, units = "px")
}

create_unequal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group, "all altitudes")
create_unequal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "high",], "high altitudes")
create_unequal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "mid",], "mid altitudes")
create_unequal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "low",], "low altitudes")

create_equal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group, "all altitudes")
create_equal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "high",], "high altitudes")
create_equal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "mid",], "mid altitudes")
create_equal_spaced_mean_fraction_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "low",], "low altitudes")

create_unequal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group, "all altitudes")
create_unequal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "high",], "high altitudes")
create_unequal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "mid",], "mid altitudes")
create_unequal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "low",], "low altitudes")

create_equal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group, "all altitudes")
create_equal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "high",], "high altitudes")
create_equal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "mid",], "mid altitudes")
create_equal_spaced_mean_number_of_plants_in_functional_groups_plot(data_functional_group[data_functional_group$altitude_band == "low",], "low altitudes")
```

Do red list analysis:
```{r}
title <- "all altitude"

analyze_red_list_data <- function(data, title) {
  tmp <- data |> 
    dplyr::select(year, global_id, CAT) |>
    group_by(year, global_id) |>
    summarise(
      n_lc = sum(CAT == 'LC'),
      n_nt = sum(CAT == 'NT'),
      n_vu = sum(CAT == 'VU'),
      n_cr = sum(CAT == 'CR'),
      n_en = sum(CAT == 'EN')
    )
  
  plot <- pivot_longer(tmp, c("n_nt", "n_vu", "n_cr", "n_en")) |> ggplot(aes(fill=name, y=value, x=year)) + 
      geom_boxplot() +
      ggtitle(paste("number of plants per plot in red list categories at ", title))
  ggsave(paste("../Daten/Results/RedList/number_of_plants_per_plot_in_red_list_cat_at", title, ".png"), plot = plot)
  
  
  tmp <- data |> 
    dplyr::select(year, global_id, CAT) |>
    group_by(year) |>
    summarise(
      n_lc = sum(CAT == 'LC') / length(unique(global_id)),
      n_nt = sum(CAT == 'NT') / length(unique(global_id)),
      n_vu = sum(CAT == 'VU') / length(unique(global_id)),
      n_cr = sum(CAT == 'CR') / length(unique(global_id)),
      n_en = sum(CAT == 'EN') / length(unique(global_id))
    )
  
  plot <-  pivot_longer(tmp, c("n_nt", "n_vu", "n_cr", "n_en")) |> ggplot(aes(fill=name, y=value, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("number of red list plant observation per plot at ", title))
  ggsave(paste("../Daten/Results/RedList/number_of_plants_per_plot_in_red_list_cat_stacked_at", title, ".png"), plot = plot)
  
  tmp <- data |> 
    dplyr::select(year, global_id, CAT) |>
    group_by(year, global_id) |>
    summarise(
      p_lc = 100 * sum(CAT == 'LC') / length(global_id),
      p_nt = 100 * sum(CAT == 'NT') / length(global_id),
      p_vu = 100 * sum(CAT == 'VU') / length(global_id),
      p_cr = 100 * sum(CAT == 'CR') / length(global_id),
      p_en = 100 * sum(CAT == 'EN') / length(global_id)
    )
  
  plot <- pivot_longer(tmp, c("p_nt", "p_vu", "p_cr", "p_en")) |> ggplot(aes(fill=name, y=value, x=year)) + 
      geom_boxplot() +
      ggtitle(paste("% of plants on each plot in red list cat at ", title))
  ggsave(paste("../Daten/Results/RedList/percent_plants_in_plots_in_red_list_cat_at", title, ".png"), plot = plot)
  
  
  tmp <- tmp |> 
    group_by(year) |>
    summarise(
      p_lc = mean(p_lc),
      p_nt = mean(p_nt),
      p_vu = mean(p_vu),
      p_cr = mean(p_cr),
      p_en = mean(p_en)
    ) 
  
  tmp <- pivot_longer(tmp, c("p_nt", "p_vu", "p_cr", "p_en"))
  
  plot <- ggplot(tmp, aes(fill=name, y=value, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("% of plants on plots in red list cat at ", title))
  ggsave(paste("../Daten/Results/RedList/percent_plants_in_plots_in_red_list_cat_stacked_at", title, ".png"), plot = plot)
  
  
  data <- distinct(data, year, `Species_unified_2004-23`, CAT)
  tmp <- data |> 
    dplyr::select(year, CAT) |>
    group_by(year) |>
    summarise(
      n_lc = sum(CAT == 'LC'),
      n_nt = sum(CAT == 'NT'),
      n_vu = sum(CAT == 'VU'),
      n_cr = sum(CAT == 'CR'),
      n_en = sum(CAT == 'EN')
    )
  
  tmp <- pivot_longer(tmp, c("n_nt", "n_vu", "n_cr", "n_en"))
  
  plot <- ggplot(tmp, aes(fill=name, y=value, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("total number of plants species in red list category at ", title))
  ggsave(paste("../Daten/Results/RedList/total_number_of_red_list_species_by_year_at_", title, ".png"), plot = plot)
}

red_list_data <- complete_data[!is.na(complete_data$CAT),]
data_red_list <- distinct(red_list_data, year, global_id, `Species_unified_2004-23`, altitude_band, CAT)

analyze_red_list_data(data_red_list, "all altitudes")
analyze_red_list_data(data_red_list[data_red_list$altitude_band == "high",], "high altitudes")
analyze_red_list_data(data_red_list[data_red_list$altitude_band == "mid",], "mid altitudes")
analyze_red_list_data(data_red_list[data_red_list$altitude_band == "low",], "low altitudes")
```

Generate .csv file for red-list and neophyte species:
```{r}

tmp <- complete_data |> 
  group_by(year, `Species_unified_2004-23`) |>
  summarise(n_total = length(altitude_band),
            n_high = sum(altitude_band == "high"),
            n_mid = sum(altitude_band == "mid"),
            n_low = sum(altitude_band == "low"))

tmp[is.na(tmp$`Species_unified_2004-23`),]
complete_data[is.na(complete_data$`Species_unified_2004-23`),]

tmp[tmp$`Species_unified_2004-23` == "Abies alba",]
tmp <- pivot_wider(tmp, names_from = c("year"), values_from = c("n_total", "n_high", "n_mid", "n_low"))
tmp[is.na(tmp)] <- 0
tmp <- left_join(tmp, red_list, join_by(`Species_unified_2004-23` == `Scientific name`))
tmp <- left_join(tmp, taxon_conversion, join_by(`Species_unified_2004-23` == infoflora_taxon))
tmp <- left_join(tmp, check_list, join_by(check_list == Taxonname))

tmp$n_total_all_years <- tmp$n_total_goedicke + tmp$n_total_2004 + tmp$n_total_2005 + tmp$n_total_2006 + tmp$n_total_2007 + tmp$n_total_2013 + tmp$n_total_2023 + tmp$n_total_climax

tmp <- distinct(tmp, `Species_unified_2004-23`, n_total_all_years, n_total_goedicke, n_total_2004, n_total_2005, n_total_2006, n_total_2007, n_total_2013, n_total_2023, n_total_climax, n_high_goedicke, n_high_2004, n_high_2005, n_high_2006, n_high_2007, n_high_2013, n_high_2023, n_high_climax, n_mid_goedicke, n_mid_2004, n_mid_2005, n_mid_2006, n_mid_2007, n_mid_2013, n_mid_2023, n_mid_climax, n_low_goedicke, n_low_2004, n_low_2005, n_low_2006, n_low_2007, n_low_2013, n_low_2023, n_low_climax, CAT, neophyt, `Indigenat CH`)
write_csv(tmp, "../Daten/Results/RedList/red_list_and_neophyte_summary.csv")
```



Do neophyte analysis:
```{r}
title <- "all altitude"

complete_data$neophyt[is.na(complete_data$neophyt)] <- F
unique(complete_data$`Species_unified_2004-23`[complete_data$neophyt])
complete_data[is.na(complete_data$`Species_unified_2004-23`),]
complete_data[is.na(complete_data$neophyt),]

data <- complete_data

analyze_neophyte_data <- function(data, title) {
  tmp <- data |> 
    dplyr::select(year, global_id, neophyt) |>
    group_by(year, global_id) |>
    summarise(
      n_neophyte = sum(neophyt)
    )
  
  plot <- tmp |> ggplot(aes(y=n_neophyte, x=year)) + 
      geom_boxplot() +
      ggtitle(paste("number of neophyte species per plot at ", title))
  ggsave(paste("../Daten/Results/Neophyte/number_of_neophytes_per_plot_at", title, ".png"), plot = plot)
  
  
  tmp <- data |> 
    dplyr::select(year, global_id, neophyt) |>
    group_by(year) |>
    summarise(
      n_neophyte = sum(neophyt) / length(unique(global_id))
    )
  
  plot <-  tmp |> ggplot(aes(y=n_neophyte, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("number of neophyte observations per plot at ", title))
  ggsave(paste("../Daten/Results/Neophyte/number_of_neophytes_per_plot_barplot_", title, ".png"), plot = plot)
  
  tmp <- data |> 
    dplyr::select(year, global_id, neophyt) |>
    group_by(year, global_id) |>
    summarise(
      p_neophyte = 100 * sum(neophyt) / length(global_id)
    )
  
  plot <- tmp |> ggplot(aes(y=p_neophyte, x=year)) + 
      geom_boxplot() +
      ggtitle(paste("% of neophytes on each plot at ", title))
  ggsave(paste("../Daten/Results/Neophyte/percent_neophytes_on_plots_at", title, ".png"), plot = plot)
  
  
  tmp <- tmp |> 
    group_by(year) |>
    summarise(
      p_neophyte = mean(p_neophyte)
    )
  
  plot <- ggplot(tmp, aes(y=p_neophyte, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("% of neophytes on plots at ", title))
  ggsave(paste("../Daten/Results/Neophyte/percent_neophytes_in_plots_at", title, ".png"), plot = plot)
  
  
  data <- distinct(data, year, `Species_unified_2004-23`, neophyt)
  tmp <- data |> 
    dplyr::select(year, neophyt) |>
    group_by(year) |>
    summarise(
      n_neophyte = sum(neophyt)
    )
  
  plot <- ggplot(tmp, aes(y=n_neophyte, x=year)) + 
      geom_bar(position="stack", stat="identity") +
      ggtitle(paste("total number of neophyte species at ", title))
  ggsave(paste("../Daten/Results/Neophyte/total_number_of_neophyte_species_by_year_at_", title, ".png"), plot = plot)
}

data_neophyte <- distinct(complete_data, year, global_id, `Species_unified_2004-23`, altitude_band, neophyt)

analyze_neophyte_data(data_neophyte, "all altitudes")
analyze_neophyte_data(data_neophyte[data_neophyte$altitude_band == "high",], "high altitudes")
analyze_neophyte_data(data_neophyte[data_neophyte$altitude_band == "mid",], "mid altitudes")
analyze_neophyte_data(data_neophyte[data_neophyte$altitude_band == "low",], "low altitudes")
```
